+++ src/chanjo2/meta/handle_report_contents.py
@@ -208,7 +208,7 @@ def get_gene_overview_coverage_stats(form_data: GeneReportForm, session: Session
             threshold_levels=form_data.completeness_thresholds
         ),
         "interval_type": form_data.interval_type.value,
-        "samples_coverage_stats_by_interval": {},
+        "interval_coverage_stats": {},
     }
 
     set_samples_coverage_files(session=session, samples=form_data.samples)
@@ -228,11 +228,27 @@ def get_gene_overview_coverage_stats(form_data: GeneReportForm, session: Session
     )
     exons_intervals = set_sql_intervals(db=session, interval_type=SQLExon, genes=[gene])
     sql_intervals = transcripts_intervals + exons_intervals
-    gene_stats["samples_coverage_stats_by_interval"] = get_gene_overview_stats(
+
+    samples_coverage_by_interval = get_gene_overview_stats(
         sql_intervals=sql_intervals,
         samples=form_data.samples,
         completeness_thresholds=form_data.completeness_thresholds,
     )
+
+    for sql_interval in sql_intervals:
+        if type(sql_interval) == SQLTranscript:
+            gene_stats["interval_coverage_stats"][sql_interval.ensembl_id] = {
+                "interval_type": "transcript",
+                "mane_select": sql_interval.refseq_mane_select,
+                "mane_plus_clinical": sql_interval.refseq_mane_plus_clinical,
+                "stats": samples_coverage_by_interval[sql_interval.ensembl_id],
+            }
+        else:
+            gene_stats["interval_coverage_stats"][sql_interval.ensembl_id] = {
+                "interval_type": "exon",
+                "stats": samples_coverage_by_interval[sql_interval.ensembl_id],
+            }
+
     return gene_stats